+++ pkg/analyzer/CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 13.2.0-dev
 
 * Allow package_config `3.x.x`.
-* Deprecate `FormalParameterElement.formalParameters` and `typeParameters`.
-  Use `FormalParameterElement.type` instead.
+* Deprecate `FormalParameterElement.formalParameters` and `typeParameters`. Use `FormalParameterElement.type` instead.
+* Added `isComplete` to `ConstructorDeclaration`, `FunctionDeclaration`, and `MethodDeclaration`.
+* Deprecated `MethodDeclaration.isAbstract`. Use `!isComplete` instead.
 
 ## 13.1.0